refactor: middleware 파일에 token.js 파일 생성 및 todos.js 파일에 헤더 토큰값 요청 첨가 (임시버전)

+++ backend/routes/todos.js
@@ -2,11 +2,19 @@ const express = require("express");
 const router = express.Router();
 const db = require("../datebase/db");
 const { StatusCodes } = require("http-status-codes");
+const { extractUserIdFromToken } = require("../middleware/token");
 
-router.get("/", (_, res) => {
-  const sql = `SELECT * FROM todos`;
+router.get("/", (req, res) => {
+  const userId = extractUserIdFromToken(req.headers.authorization);
+  const sql = `SELECT * FROM todos WHERE user_id = ?`;
+  console.log("Request Headers:", req.headers);
+  console.log(req.headers.authorization);
 
-  db.all(sql, [], (err, rows) => {
+  if(!userId) {
+    return res.status(StatusCodes.UNAUTHORIZED).json({ message: "userId is missing"})
+  }
+
+  db.all(sql, [userId], (err, rows) => {
     if (err) {
       return res
         .status(StatusCodes.INTERNAL_SERVER_ERROR)
@@ -17,8 +25,13 @@ router.get("/", (_, res) => {
 });
 
 router.post("/", (req, res) => {
+  const userId = extractUserIdFromToken(req.headers.authorization);
   const { title } = req.body;
-  const sql = `INSERT INTO todos (title) VALUES (?)`;
+  const sql = `INSERT INTO todos (title, user_id) VALUES (?, ?)`;
+
+  if(!userId) {
+    return res.status(StatusCodes.UNAUTHORIZED).json({ message: "userId is missing"})
+  }
 
   if (!title || !title.trim()) {
     return res
@@ -26,7 +39,7 @@ router.post("/", (req, res) => {
       .json({ message: "Invalid input format" });
   }
 
-  db.run(sql, [title], (err) => {
+  db.run(sql, [title, userId], (err) => {
     if (err) {
       return res
         .status(StatusCodes.INTERNAL_SERVER_ERROR)
@@ -39,11 +52,16 @@ router.post("/", (req, res) => {
 });
 
 router.patch("/:id", (req, res) => {
+  const userId = extractUserIdFromToken(req.headers.authorization);
   const { id } = req.params;
   const { done } = req.body;
-  const sql = `UPDATE todos SET done = ? WHERE id = ?;`;
+  const sql = `UPDATE todos SET done = ? WHERE id = ?`;
 
-  db.get(`SELECT * FROM todos WHERE id = ?;`, [id], (err, row) => {
+  if(!userId) {
+    return res.status(StatusCodes.UNAUTHORIZED).json({ message: "userId is missing"})
+  }
+
+  db.get(`SELECT * FROM todos WHERE id = ? AND user_id = ?`, [id, userId], (err, row) => {
     if (err) {
       return res
         .status(StatusCodes.INTERNAL_SERVER_ERROR)
@@ -68,9 +86,14 @@ router.patch("/:id", (req, res) => {
 });
 
 router.put("/:id", (req, res) => {
+  const userId = extractUserIdFromToken(req.headers.authorization);
   const { id } = req.params;
   const { title } = req.body;
-  const sql = `UPDATE todos SET title = ? WHERE id = ?;`;
+  const sql = `UPDATE todos SET title = ? WHERE id = ? AND user_id = ?`;
+
+  if(!userId) {
+    return res.status(StatusCodes.UNAUTHORIZED).json({ message: "userId is missing"})
+  }
 
   if (!title || !title.trim()) {
     return res
@@ -78,7 +101,7 @@ router.put("/:id", (req, res) => {
       .json({ message: "Invalid input format" });
   }
 
-  db.get(`SELECT * FROM todos WHERE id = ?;`, [id], (err, row) => {
+  db.get(`SELECT * FROM todos WHERE id = ? AND user_id = ?`, [id, userId], (err, row) => {
     if (err) {
       return res
         .status(StatusCodes.INTERNAL_SERVER_ERROR)
@@ -91,7 +114,7 @@ router.put("/:id", (req, res) => {
         .json({ message: "TO DO id not found" });
     }
 
-    db.run(sql, [title, id], function (err) {
+    db.run(sql, [title, id, userId], function (err) {
       if (err) {
         return res
           .status(StatusCodes.INTERNAL_SERVER_ERROR)
@@ -104,10 +127,15 @@ router.put("/:id", (req, res) => {
 });
 
 router.delete("/:id", (req, res) => {
+  const userId = extractUserIdFromToken(req.headers.authorization);
   const { id } = req.params;
-  const sql = `DELETE FROM todos WHERE id = ?;`;
+  const sql = `DELETE FROM todos WHERE id = ? AND user_id = ?`;
 
-  db.get(`SELECT * FROM todos WHERE id = ?;`, [id], (err, row) => {
+  if(!userId) {
+    return res.status(StatusCodes.UNAUTHORIZED).json({ message: "userId is missing"})
+  }
+
+  db.get(`SELECT * FROM todos WHERE id = ? AND user_id = ?`, [id, userId], (err, row) => {
     if (err) {
       return res
         .status(StatusCodes.INTERNAL_SERVER_ERROR)
@@ -120,7 +148,7 @@ router.delete("/:id", (req, res) => {
         .json({ message: "TO DO id not found" });
     }
 
-    db.run(sql, [id], function (err) {
+    db.run(sql, [id, userId], function (err) {
       if (err) {
         return res
           .status(StatusCodes.INTERNAL_SERVER_ERROR)
